Create a special saved enpoint that accepts a fileID
The saved endpoint will accept the file id as a parameter and
return the same page that would be given by normally going to
the page enpoint with a url parameter instead.

The javascript frontend must have a way of uniquely identifying
which file version is being read by the user. This has yet to be
implemented.

There is also now a page where users can look at their saved pages
and access them directly without going through the splash page
and looking the url up manually.